+++ bcos-crypto/interfaces/crypto/Concepts.h
@@ -30,7 +30,7 @@ concept TrivialValue = std::is_trivial_v<std::remove_cvref_t<Value>>;
 
 #if (defined __clang__) && (__clang_major__ < 15)
 template <class Range>
-concept TrivialRange = std::ranges::random_access_range<std::remove_cvref_t<Range>> &&
+concept TrivialRange = std::ranges::range<std::remove_cvref_t<Range>> &&
     std::is_trivial_v<std::remove_cvref_t<std::ranges::range_value_t<Range>>>;
 #else
 template <class Range>
