+++ bcos-crypto/interfaces/crypto/hasher/Hasher.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "../Common.h"
+#include "../Concepts.h"
 #include <boost/throw_exception.hpp>
 #include <iterator>
 #include <ranges>
