+++ demo/perf_demo.cpp
@@ -18,7 +18,7 @@
  * @date 2021.04.07
  * @author yujiechen
  */
-#include <bcos-crypto/crypto2/hasher/OpenSSLHasher.h>
+#include <bcos-crypto/interfaces/crypto/hasher/OpenSSLHasher.h>
 #include <bcos-crypto/encrypt/AESCrypto.h>
 #include <bcos-crypto/encrypt/SM4Crypto.h>
 #include <bcos-crypto/hash/Keccak256.h>
